+++ frontend/src/types.ts
@@ -6,3 +6,10 @@ export type Item = {
   currency_code: string;
   emission_factor_name?: string;
 };
+
+export type CreateLineItemRequest = {
+  description: string;
+  quantity: number;
+  unit_amount: number;
+  currency_code: string;
+};
